Move PixelFormat to pixels folder
